% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_package.R
\name{multitools-package}
\alias{multitools-package}
\title{multitools: Tools for Contributing Packages to R-multiverse}
\description{
R-multiverse is a community-curated collection of
R package releases, powered by R-universe. The \code{multitools} package
has tools for maintainers of packages in R-multiverse.
}
\concept{help}
