% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_package.R
\name{releases-package}
\alias{releases-package}
\title{releases: Tools for Package Releases}
\description{
R-releases is a community-curated,
maintainer-driven repository of R package releases
based on R-universe. The \code{releases} package
has tools for maintainers of packages in R-releases.
}
\concept{help}
